changing data type to float

+++ igf_data/utils/project_data_display_utils.py
@@ -75,12 +75,12 @@ def convert_project_data_gviz_data(input_data,output_file):
     if len(list(seqrun_set)) >1:
         description.update({"total_read":("number", "Total Reads")})            # add total read column for samples with multiple runs
         intermediate_data['total_read']=intermediate_data['total_read'].\
-                                        astype(int)                             # convert column to number
+                                        astype(float)                           # convert column to number
 
     for run in list(seqrun_set):
         description.update({run:("number",run)})                                # add seqrun columns
         intermediate_data[run]=intermediate_data[run].\
-                               astype(int)                                      # convert column to number
+                               astype(float)                                    # convert column to number
 
     intermediate_data=intermediate_data.to_dict(orient='region')                # convert data frame to json
     data_table = gviz_api.DataTable(description)                                # load description to gviz api
